requirements.txt
numpy
pandas
scikit-learn
seaborn
jupyter
matplotlib
plotly
missingno
imbalanced-learn
dune-client
python-dotenv
joblib
xgboost
Flask
gunicorn
nbformat
requests
fastapi
uvicorn
aiohttp
pydantic